# Refactor EGraph and UF internals for clarity and performance